+++ OpenClash/OpenClash(mihomo-smart).sh
@@ -51,7 +51,7 @@ dns:
   enable: true
   listen: 0.0.0.0:7874
   ipv6: false
-  prefer-h3: true
+  prefer-h3: false
   enhanced-mode: fake-ip
   fake-ip-range: 198.18.0.1/16
   fake-ip-filter: